requirements.txt
# AI Trading Bot - Production Dependencies
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
aiohttp==3.11.11

# Data & Analysis
pandas>=2.2.0,<3
numpy>=1.26.0,<2
ta==0.11.0

# Machine Learning (use flexible version for Raspberry Pi / ARM; 2.16.x on x86, 2.20.x on Pi)
tensorflow>=2.16.2,<2.21
scikit-learn==1.6.0

# Database
aiosqlite==0.20.0

# Security
cryptography==44.0.0
python-dotenv==1.0.1

# Monitoring & Logging
structlog==24.4.0
rich==13.9.4

# Configuration
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
ruamel.yaml==0.18.6

# Async Utilities
asyncio-throttle==1.0.2

# Telegram Bot (optional)
python-telegram-bot==21.9

# Billing (Stripe)
stripe>=8.0.0,<10

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Type checking
mypy==1.14.1
